requirements.txt
Flask
flask_restx
flask_cors
sklearn
nltk
gunicorn
google-cloud-datastore